requirements.txt
pandas
google-api-python-client
google.cloud.language
plotext